refactor: move free_tcp_ipv4_port() to common
The backends module, along with the llama_cpp and vllm modules,
have a circular dependencies, which is considered an anti-pattern.

Move free_tcp_ipv4_port() to reduce the circular dependencies.
