+++ src/instructlab/model/backends/backends.py
@@ -12,7 +12,6 @@ import multiprocessing
 import os
 import pathlib
 import signal
-import socket
 import struct
 import subprocess
 import sys
@@ -398,19 +397,6 @@ def wait_for_stable_vram_cuda(timeout: int) -> Tuple[bool, bool]:
             torch.cuda.empty_cache()
         except Exception as e:  # pylint: disable=broad-exception-caught
             logger.debug("Could not free cuda cache: %s", e)
-
-
-def free_tcp_ipv4_port(host: str) -> int:
-    """Ask the OS for a random, ephemeral, and bindable TCP/IPv4 port
-
-    Note: The idea of finding a free port is bad design and subject to
-    race conditions. Instead vLLM and llama-cpp should accept port 0 and
-    have an API to return the actual listening port. Or they should be able
-    to use an existing socket like a systemd socket activation service.
-    """
-    with socket.socket(socket.AF_INET, socket.SOCK_STREAM) as s:
-        s.bind((host, 0))
-        return int(s.getsockname()[-1])
 
 
 def is_temp_server_running():
