+++ src/instructlab/model/backends/vllm.py
@@ -18,18 +18,14 @@ import httpx
 # Local
 from ...client import check_api_base
 from ...configuration import get_api_base
-from .backends import (
-    BackendServer,
-    free_tcp_ipv4_port,
-    safe_close_all,
-    shutdown_process,
-)
+from .backends import BackendServer, safe_close_all, shutdown_process
 from .common import (
     CHAT_TEMPLATE_AUTO,
     CHAT_TEMPLATE_TOKENIZER,
     VLLM,
     Closeable,
     ServerException,
+    free_tcp_ipv4_port,
     get_model_template,
     verify_template_exists,
 )
